Rename scene files to use lowercase strings
