+++ _tests_/menuscene.test.js
@@ -1,0 +1,1 @@
+import PlayerScene from '../src/';
