+++ src/scenes/level-one.js
@@ -1,40 +1,52 @@
-import PlayerScene from './PlayerScene';
+import PlayerScene from './player';
+import Helpers from '../helpers/utils';
 
-class Level2Scene extends PlayerScene {
+class Level1Scene extends PlayerScene {
   constructor() {
-    super('Level2Scene');
+    super('Level1Scene');
   }
 
   preload() {
-    this.load.image('level_two_platform', 'assets/images/platform_2.png');
+    this.load.image('level_one_platform', 'assets/images/platform.png');
   }
 
   create() {
-    this.physics.resume();
+    const bgCoordinates = [
+      { x: 100, y: 100 },
+      { x: 500, y: 150 },
+      { x: 800, y: 100 },
+      { x: 300, y: 300 },
+      { x: 100, y: 500 },
+      { x: 900, y: 300 },
+      { x: 600, y: 400 },
+    ];
 
     const platformPositions = [
-      { x: 92, y: 580 },
-      { x: 272, y: 580 },
-      { x: 452, y: 580 },
-      { x: 632, y: 580 },
-      { x: 812, y: 580 },
-      { x: 992, y: 580 },
+      { x: 158, y: 580 },
+      { x: 350, y: 580 },
+      { x: 500, y: 580 },
+      { x: 800, y: 580 },
+      { x: 1000, y: 580 },
     ];
+
+    const nameInput = document.getElementById('nameInput');
+    const submitBtn = document.getElementById('submitBtn');
+
+    Helpers.hideElement(nameInput);
+    Helpers.hideElement(submitBtn);
 
     this.gameOnSound = this.sound.add('game-music', { loop: true });
     this.gameOverSound = this.sound.add('game_over_sound', { loop: false });
     this.winSound = this.sound.add('win-sound');
     this.gameOnSound.play();
 
-    this.backgroundImage = this.add.image(0, 0, 'level_two_bg').setOrigin(0, 0);
-    this.backgroundImage.displayWidth = this.sys.canvas.width;
-    this.backgroundImage.displayHeight = this.sys.canvas.height;
-    this.score = 3000;
-
     this.scoreText = this.add.text(10, 25, `Score: ${this.score}`, {
+      fill: '#37c3be',
       fontSize: 23,
       fontWeight: 900,
     });
+
+    bgCoordinates.map(bgCoord => this.add.image(bgCoord.x, bgCoord.y, 'level1_bg'));
 
     this.gameState.character = this.physics.add
       .sprite(225, 440, 'character')
@@ -43,25 +55,28 @@ class Level2Scene extends PlayerScene {
 
     const platforms = this.physics.add.staticGroup();
     this.physics.add.collider(this.gameState.character, platforms);
-    this.gameState.character.setCollideWorldBounds(true);
 
     platformPositions.map(plat => platforms
-      .create(plat.x, plat.y, 'level_two_platform')
-      .setScale(0.9));
+      .create(plat.x, plat.y, 'level_one_platform')
+      .setScale(0.7));
+
+    this.gameState.character.setCollideWorldBounds(true);
+    this.physics.add.collider(this.gameState.character, platforms);
 
     const bugs = this.physics.add.group();
-    const thebugs = ['green_bug', 'yellow_bug'];
+    const bugList = ['green_bug', 'green_bug', 'green_bug'];
 
+    // generate & display bugs randomly on the game screen.
     function bugGenerator() {
       const xCoordinate = Math.random() * 1000;
-      const randomBug = thebugs[Math.floor(Math.random() * 2)];
+      const randomBug = bugList[Math.floor(Math.random() * 3)];
 
       bugs.create(xCoordinate, 10, randomBug);
     }
 
     const bugObject = {
       callback: bugGenerator,
-      delay: 90,
+      delay: 150,
       callbackScope: this,
       loop: true,
     };
@@ -98,7 +113,6 @@ class Level2Scene extends PlayerScene {
       this.bugGeneratorLoop.destroy();
       this.gameOnSound.stop();
       this.gameOverSound.play();
-
       this.physics.pause();
       this.registry.set('playerScore', this.score);
       this.scene.start('GameOverScene');
@@ -106,11 +120,12 @@ class Level2Scene extends PlayerScene {
   }
 
   update() {
-    if (this.score === 10000) {
-      this.startNextLevel('Level3Scene');
+    if (this.score === 3000) {
+      this.startNextLevel('Level2Scene');
     }
+
     this.movePlayer();
   }
 }
 
-export default Level2Scene;
+export default Level1Scene;
