+++ src/scenes/level-two.js
@@ -1,15 +1,17 @@
-import PlayerScene from './PlayerScene';
+import PlayerScene from './player';
 
-class Level3Scene extends PlayerScene {
+class Level2Scene extends PlayerScene {
   constructor() {
-    super('Level3Scene');
+    super('Level2Scene');
   }
 
   preload() {
-    this.load.image('level_three_platform', 'assets/images/platform_2.png');
+    this.load.image('level_two_platform', 'assets/images/platform_2.png');
   }
 
   create() {
+    this.physics.resume();
+
     const platformPositions = [
       { x: 92, y: 580 },
       { x: 272, y: 580 },
@@ -21,12 +23,13 @@ class Level3Scene extends PlayerScene {
 
     this.gameOnSound = this.sound.add('game-music', { loop: true });
     this.gameOverSound = this.sound.add('game_over_sound', { loop: false });
+    this.winSound = this.sound.add('win-sound');
     this.gameOnSound.play();
 
-    this.background = this.add.image(0, 0, 'level_three_bg').setOrigin(0, 0);
-    this.background.displayWidth = this.sys.canvas.width;
-    this.background.displayHeight = this.sys.canvas.height;
-    this.score = 10000;
+    this.backgroundImage = this.add.image(0, 0, 'level_two_bg').setOrigin(0, 0);
+    this.backgroundImage.displayWidth = this.sys.canvas.width;
+    this.backgroundImage.displayHeight = this.sys.canvas.height;
+    this.score = 3000;
 
     this.scoreText = this.add.text(10, 25, `Score: ${this.score}`, {
       fontSize: 23,
@@ -37,30 +40,34 @@ class Level3Scene extends PlayerScene {
       .sprite(225, 440, 'character')
       .setScale(0.6);
     this.gameState.cursors = this.input.keyboard.createCursorKeys();
+
     const platforms = this.physics.add.staticGroup();
     this.physics.add.collider(this.gameState.character, platforms);
     this.gameState.character.setCollideWorldBounds(true);
 
-    platformPositions.map(plat => platforms.create(plat.x, plat.y, 'level_three_platform'));
+    platformPositions.map(plat => platforms
+      .create(plat.x, plat.y, 'level_two_platform')
+      .setScale(0.9));
 
     const bugs = this.physics.add.group();
-    const theBugList = ['green_bug', 'yellow_bug', 'red_bug'];
+    const thebugs = ['green_bug', 'yellow_bug'];
 
     function bugGenerator() {
       const xCoordinate = Math.random() * 1000;
-      const randomBug = theBugList[Math.floor(Math.random() * 3)];
+      const randomBug = thebugs[Math.floor(Math.random() * 2)];
 
       bugs.create(xCoordinate, 10, randomBug);
     }
 
     const bugObject = {
       callback: bugGenerator,
-      delay: 50,
+      delay: 90,
       callbackScope: this,
       loop: true,
     };
 
     this.bugGeneratorLoop = this.time.addEvent(bugObject);
+
     // create animations on the character sprite
     this.anims.create({
       key: 'run',
@@ -99,8 +106,11 @@ class Level3Scene extends PlayerScene {
   }
 
   update() {
+    if (this.score === 10000) {
+      this.startNextLevel('Level3Scene');
+    }
     this.movePlayer();
   }
 }
 
-export default Level3Scene;
+export default Level2Scene;
