+++ src/scenes/player.js
